Refactor manifest parser classes
